Move session functions to helpers.js
